# backend/internal/transport/rest/apidocs/assets/openapi.yaml
openapi: 3.0.3
info:
  title: SCYG Blog Content API
  version: 1.0.0
  description: "SCYG 博客内容 API，提供公开读取和受授权保护的写入。"
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [Auth]
      security: []
      summary: "登录并获取访问令牌"
      description: "使用用户名和密码验证用户身份，成功后返回短期 Bearer JWT。"
      requestBody:
        description: "用户登录凭据。"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: "登录成功。"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/articles:
    get:
      operationId: listArticles
      security: []
      tags: [Articles]
      summary: "查询已发布文章"
      description: "分页返回状态为 `2`（已发布）且未删除的文章，可按文章类型、标签和关键词筛选。"
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/ArticleTypeFilter'
        - $ref: '#/components/parameters/TagFilter'
        - $ref: '#/components/parameters/SearchFilter'
      responses:
        '200':
          description: "请求成功，返回当前资源或分页响应。"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/articles/{articleId}:
    parameters:
      - $ref: '#/components/parameters/ArticleID'
    get:
      operationId: getArticle
      security: []
      tags: [Articles]
      summary: "获取已发布文章"
      description: "仅在文章状态为 `2`（已发布）且未删除时返回。"
      responses:
        '200':
          description: "请求成功，返回当前资源或分页响应。"
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/article-types:
    get:
      operationId: listPublicArticleTypes
      security: []
      tags: [ArticleTypes]
      summary: "查询公开文章类型"
      description: "分页返回有公开文章的未删除文章类型，按公开文章数量降序、标识升序返回。"
      parameters:
        - $ref: '#/components/parameters/PublicPage'
        - $ref: '#/components/parameters/PublicPageSize'
        - $ref: '#/components/parameters/PublicSearchFilter'
      responses:
        '200':
          description: "请求成功。"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicArticleTypeList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/article-types/{articleTypeId}:
    parameters:
      - $ref: '#/components/parameters/ArticleTypeID'
    get:
      operationId: getPublicArticleType
      security: []
      tags: [ArticleTypes]
      summary: "获取公开文章类型"
      description: "返回有公开文章的未删除文章类型。"
      responses:
        '200':
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          description: "请求成功。"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicArticleType'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/tags:
    get:
      operationId: listPublicTags
      security: []
      tags: [Tags]
      summary: "查询公开标签"
      description: "分页返回关联公开文章的未删除标签，按公开文章数量降序、标识升序返回。"
      parameters:
        - $ref: '#/components/parameters/PublicPage'
        - $ref: '#/components/parameters/PublicPageSize'
        - $ref: '#/components/parameters/PublicSearchFilter'
      responses:
        '200':
          description: "请求成功。"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicTagList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/tags/{tagId}:
    parameters:
      - $ref: '#/components/parameters/TagID'
    get:
      operationId: getPublicTag
      security: []
      tags: [Tags]
      summary: "获取公开标签"
      description: "返回关联公开文章的未删除标签。"
      responses:
        '200':
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          description: "请求成功。"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicTag'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /media/article-images/{storageKey}:
    parameters:
      - $ref: '#/components/parameters/ArticleImageStorageKey'
    get:
      operationId: getArticleImageMedia
      security: []
      tags: [ArticleImages]
      summary: "读取正文图片媒体"
      description: "按受控存储键读取 JPEG 或 PNG 二进制；实体标签未变化时返回 304。"
      responses:
        '200':
          description: "读取成功，返回原始图片二进制。"
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            image/jpeg:
              schema:
                description: "JPEG 图片二进制。"
                type: string
                format: binary
            image/png:
              schema:
                description: "PNG 图片二进制。"
                type: string
                format: binary
        '304':
          description: "实体标签未变化，不返回响应体。"
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/runs/{runId}:
    get:
      operationId: agentGet
      tags: [Agent]
      summary: "查询 Run 和最终结果"
      description: "使用 Blog Bearer JWT 查询当前 Run；Agent 校验 owner，Run 不存在和非 owner 均返回 404。最终结果在此响应中读取，不提供独立 result 接口。Agent 集成关闭时返回 404。"
      parameters:
        - $ref: '#/components/parameters/AgentRunID'
      responses:
        '200':
          description: "查询成功，返回统一 Run 快照和已有结果。"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRun'
        '400':
          $ref: '#/components/responses/AgentError'
        '401':
          $ref: '#/components/responses/AgentError'
        '404':
          $ref: '#/components/responses/AgentError'
        '409':
          $ref: '#/components/responses/AgentError'
        '429':
          $ref: '#/components/responses/AgentError'
        '503':
          $ref: '#/components/responses/AgentError'
        '500':
          $ref: '#/components/responses/AgentError'
        default:
          $ref: '#/components/responses/AgentError'
  /api/v1/runs/{runId}/events:
    get:
      operationId: agentEvents
      tags: [Agent]
      summary: "订阅 Run 事件流"
      description: "使用 fetch 和 ReadableStream 携带 Blog Bearer JWT 读取 SSE，不使用 Cookie、URL token 或原生 EventSource。先查询 Run 完成授权，再订阅；Agent 确认订阅建立后才提交 HTTP 200，不等待首条业务事件，建立前失败返回 Problem Details。Agent 的完整 SSE frame（含 event、id、data、注释和 heartbeat）原样转发并逐条 flush；Blog 不重建事件、不生成序号或 heartbeat，不把 gRPC metadata/trailers 转为 HTTP headers。订阅建立后连续未收到任何 frame 达 idle timeout、流结束或异常时直接关闭，不生成 stream_error。断线后查询 Run：非终态重新订阅，终态停止订阅。Agent 集成关闭时返回 404。"
      parameters:
        - $ref: '#/components/parameters/AgentRunID'
        - $ref: '#/components/parameters/AgentEventCursor'
      responses:
        '200':
          description: "订阅建立成功，持续返回原始 SSE frame。"
          content:
            text/event-stream:
              schema:
                description: "Agent 编码的完整 SSE frame 字节流。"
                type: string
        '400':
          $ref: '#/components/responses/AgentError'
        '401':
          $ref: '#/components/responses/AgentError'
        '404':
          $ref: '#/components/responses/AgentError'
        '409':
          $ref: '#/components/responses/AgentError'
        '429':
          $ref: '#/components/responses/AgentError'
        '503':
          $ref: '#/components/responses/AgentError'
        '500':
          $ref: '#/components/responses/AgentError'
        default:
          $ref: '#/components/responses/AgentError'
  /api/v1/runs/{runId}/resume:
    post:
      operationId: agentResume
      tags: [Agent]
      summary: "恢复待交互 Run"
      description: "使用 Blog Bearer JWT 提交交互决定，Agent 校验 owner、交互类型、决定内容和恢复条件，Blog 不解释 decision 或 payload。payload 省略与显式 null 分别透传；同 key 重试按成功记录重放。交互已处理后以新 key 再提交返回 409，不覆盖原决定。Run 不存在和非 owner 均返回 404。Agent 集成关闭时返回 404。"
      parameters:
        - $ref: '#/components/parameters/AgentRunID'
        - $ref: '#/components/parameters/AgentIdempotencyKey'
      requestBody:
        required: true
        description: "必须携带 application/json（允许 charset 参数）；请求体须为合法 UTF-8 JSON，最大 1 MiB，空请求体或非法 JSON 返回 400，超限返回 413，媒体类型缺失或不支持返回 415。顶层必须为 object，含非空 interactionId 和 decision。"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentResumeRequest'
      responses:
        '200':
          description: "恢复操作成功，返回最新统一 Run 快照。"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRun'
        '400':
          $ref: '#/components/responses/AgentError'
        '401':
          $ref: '#/components/responses/AgentError'
        '404':
          $ref: '#/components/responses/AgentError'
        '409':
          $ref: '#/components/responses/AgentError'
        '429':
          $ref: '#/components/responses/AgentError'
        '503':
          $ref: '#/components/responses/AgentError'
        '500':
          $ref: '#/components/responses/AgentError'
        '413':
          $ref: '#/components/responses/AgentError'
        '415':
          $ref: '#/components/responses/AgentError'
        default:
          $ref: '#/components/responses/AgentError'
  /api/v1/runs/{runId}/cancel:
    post:
      operationId: agentCancel
      tags: [Agent]
      summary: "取消 Run"
      description: "使用 Blog Bearer JWT，请求不得包含 body。Agent 设置取消栅栏后立即返回 cancelled，不再发起新的 Tool 调用；在途 Blog 写入可以完成且结果保留，但不再将 Run 推进为成功。已 cancelled 的 Run 使用新 key 再取消返回当前 Run；已 succeeded 或 failed 的 Run 返回 409。Run 不存在和非 owner 均返回 404。Agent 集成关闭时返回 404。"
      parameters:
        - $ref: '#/components/parameters/AgentRunID'
        - $ref: '#/components/parameters/AgentIdempotencyKey'
      responses:
        '200':
          description: "取消操作成功，返回最新统一 Run 快照。"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRun'
        '400':
          $ref: '#/components/responses/AgentError'
        '401':
          $ref: '#/components/responses/AgentError'
        '404':
          $ref: '#/components/responses/AgentError'
        '409':
          $ref: '#/components/responses/AgentError'
        '429':
          $ref: '#/components/responses/AgentError'
        '503':
          $ref: '#/components/responses/AgentError'
        '500':
          $ref: '#/components/responses/AgentError'
        '413':
          $ref: '#/components/responses/AgentError'
        default:
          $ref: '#/components/responses/AgentError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AgentRunID:
      description: "Agent 生成的不透明 URL-safe ASCII 标识，长度 1–128；只允许 A–Z、a–z、0–9、点、下划线、波浪号、连字符，不能等于 . 或 ..。"
      name: runId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/AgentRunID'
...
    AgentEventCursor:
      description: "可选的不透明事件游标；提供时必须为 1–256 个 ASCII 字节。Blog 不解析、不比较、不持久化，作为 after_event_id 原样传给 Agent；省略时由 Agent 使用默认订阅位置。"
      name: Last-Event-ID
      in: header
      schema:
        type: string
        minLength: 1
        maxLength: 256
        pattern: '^[\x00-\x7F]+$'
...
    ArticleImageStorageKey:
      description: "32 位小写十六进制正文图片键及 `.jpg` 或 `.png` 扩展名。"
      name: storageKey
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ArticleImageStorageKey'
    ArticleID:
      description: "文章正整数标识。"
      name: articleId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/PositiveID'
    ArticleTypeID:
      description: "文章类型正整数标识。"
      name: articleTypeId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/PositiveID'
    TagID:
      description: "标签正整数标识。"
      name: tagId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/PositiveID'
...
    Page:
      description: "从 `1` 开始的页码，默认 `1`。"
      name: page
      in: query
      schema:
        type: integer
        format: int32
        default: 1
        minimum: 1
    PageSize:
      description: "每页数量，默认 `20`，范围 `1` 至 `100`。"
      name: pageSize
      in: query
      schema:
        type: integer
        format: int32
        default: 20
        minimum: 1
        maximum: 100
    PublicPage:
      description: "从 `1` 开始的必填页码。"
      name: page
      in: query
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
    PublicPageSize:
      description: "必填的每页数量，范围 `1` 至 `100`。"
      name: pageSize
      in: query
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
    PublicSearchFilter:
      description: "按名称筛选；服务端先去除首尾空白，再进行不区分大小写的包含匹配。"
      name: q
      in: query
      schema:
        type: string
        minLength: 1
        maxLength: 200
        example: Go
    Sort:
      description: "排序：`createdAt` 创建时间升序，`-createdAt` 降序，`updatedAt` 更新时间升序，`-updatedAt` 降序，`title` 标题升序，`-title` 降序。"
      name: sort
      in: query
      schema:
        description: "`createdAt` 创建时间升序，`-createdAt` 创建时间降序，`updatedAt` 更新时间升序，`-updatedAt` 更新时间降序，`title` 标题升序，`-title` 标题降序。"
        type: string
        default: -createdAt
        enum: [createdAt, -createdAt, updatedAt, -updatedAt, title, -title]
        example: -createdAt
    ArticleTypeFilter:
      description: "按文章类型标识筛选文章，例如 `1`。"
      name: articleTypeId
      in: query
      schema:
        $ref: '#/components/schemas/PositiveID'
    TagFilter:
      description: "按标签标识筛选文章，例如 `1`。"
      name: tagId
      in: query
      schema:
        $ref: '#/components/schemas/PositiveID'
    SearchFilter:
      description: "按关键词筛选可搜索文本，例如 `Go`。"
      name: q
      in: query
      schema:
        type: string
        minLength: 1
        maxLength: 200
        example: Go
  headers:
    ETag:
      description: "强实体标签，内容为当前正整数版本，例如 `\"1\"`。"
      required: true
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
        example: '"1"'
...
  responses:
    AgentError:
      description: "RFC 9457 错误响应。HTTP 400：非法 ID、key、JSON 或请求结构，或 Agent INVALID_ARGUMENT；401：JWT 无效或 UNAUTHENTICATED；404：集成关闭、NOT_FOUND 或 PERMISSION_DENIED（含非 owner）；409：ALREADY_EXISTS、ABORTED 或 FAILED_PRECONDITION；413：body 超过 1 MiB；415：创建或 Resume 的 Content-Type 缺失或不是 application/json；429：RESOURCE_EXHAUSTED；503：UNAVAILABLE 或 DEADLINE_EXCEEDED；其他 Agent 错误为 500。建立 SSE 之前同样适用，提交 200 后仅关闭流。Agent 提供的公开 code、message 已脱敏，缺失时使用本地通用消息；不透传 Provider 原始错误、数据库信息、密钥或敏感 Tool 参数。"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    BadRequest:
      description: "HTTP `400`，请求参数或请求体校验失败。"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: "HTTP `401`，缺少或无法验证认证凭据。"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: "HTTP `404`，资源不存在或不可公开访问。"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    InternalError:
      description: "HTTP `500`，服务发生未处理错误。"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    AgentRun:
      description: "Create、Get、Resume、Cancel 共用的 owner 授权 Run 快照，字段始终存在；无结果、失败或待处理交互时对应值为 null。result 与 failure 可以同时存在，Blog 不实现 Run 状态机。"
      type: object
      required: [runId, status, capability, recipeId, recipeVersion, createdAt, updatedAt, failure, result, pendingInteraction, streamUrl]
      properties:
        runId:
          $ref: '#/components/schemas/AgentRunID'
        status:
          description: "Run 状态：`queued` 排队，`running` 执行中，`waitingForApproval` 等待交互，`succeeded` 成功，`failed` 失败，`cancelled` 已取消；后三者为终态。"
          type: string
          enum: [queued, running, waitingForApproval, succeeded, failed, cancelled]
        capability:
          description: "路由固定能力：`search` 搜索，`write` 写作，`polish` 润色，`chat` 对话。"
          type: string
          enum: [search, write, polish, chat]
        recipeId:
          description: "Agent 选定的 Recipe 标识，创建成功起非空，客户端只读。"
          type: string
          minLength: 1
          readOnly: true
        recipeVersion:
          description: "Agent 选定的 Recipe 版本，创建成功起非空，客户端只读。"
          type: string
          minLength: 1
          readOnly: true
        createdAt:
          description: "Run 创建时间，RFC 3339 UTC 字符串。"
          type: string
          format: date-time
          pattern: 'Z$'
        updatedAt:
          description: "Run 最近更新时间，RFC 3339 UTC 字符串。"
          type: string
          format: date-time
          pattern: 'Z$'
        streamUrl:
          description: "Blog 根据 runId 生成的相对 SSE 订阅路径，例如 /api/v1/runs/run-safe-id/events。"
          type: string
          format: uri-reference
          pattern: '^/api/v1/runs/[A-Za-z0-9._~-]{1,128}/events$'
        result:
          description: "Agent 提供的任意 JSON 结果；无结果时为 null，可以与 failure 同时存在。"
          nullable: true
        failure:
          description: "Agent 提供的公开失败对象，无失败时为 null；Blog 不翻译或解释 code 和 message。"
          type: object
          nullable: true
          required: [code, message]
          properties:
            code:
              description: "Agent 提供的失败代码。"
              type: string
            message:
              description: "Agent 提供的已脱敏失败说明。"
              type: string
        pendingInteraction:
          description: "当前待处理交互，无待处理交互时为 null。"
          type: object
          nullable: true
          required: [interactionId, type, payload]
          properties:
            interactionId:
              description: "Agent 提供的不透明交互标识，Resume 时原样提交。"
              type: string
            type:
              description: "交互类型：`confirmation` 确认，`selection` 选择，`textInput` 文本输入。"
              type: string
              enum: [confirmation, selection, textInput]
            payload:
              description: "Agent 定义的任意 JSON 交互内容，可以为 null。"
              nullable: true
...
    LoginRequest:
      description: "用户名和密码登录凭据。"
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          description: "登录用户名。"
          type: string
          minLength: 1
          maxLength: 64
        password:
          description: "登录密码，不在错误响应中回显。"
          type: string
          minLength: 1
          maxLength: 256
    LoginResponse:
      description: "短期访问令牌。"
      type: object
      additionalProperties: false
      required: [accessToken, tokenType, expiresAt]
      properties:
        accessToken:
          description: "短期 Bearer 访问令牌。"
          type: string
          minLength: 1
        tokenType:
          description: "令牌类型：`Bearer` 表示 HTTP Bearer 认证。"
          type: string
          enum: [Bearer]
        expiresAt:
          description: "令牌过期时间。"
          type: string
          format: date-time
...
    PositiveID:
      description: "资源正整数标识，示例 `1`。"
      example: 1
      type: integer
      format: int64
      minimum: 1
...
    ArticleStatus:
      description: "文章状态：`1` 草稿，`2` 已发布，`3` 已归档。"
      example: 2
      type: integer
      format: int32
      enum: [1, 2, 3]
      x-enum-varnames: [Draft, Published, Archived]
...
    PublicArticleType:
      description: "公开文章类型资源。"
      type: object
      additionalProperties: false
      required: [id, name, image, articleCount]
      properties:
        id:
          $ref: '#/components/schemas/PositiveID'
        name:
          description: "文章类型名称。"
          type: string
          minLength: 1
          maxLength: 60
        image:
          description: "分类图片地址，未设置时为 `null`。"
          type: string
          maxLength: 512
          nullable: true
        articleCount:
          description: "满足公开条件的文章数量。"
          type: integer
          format: int64
          minimum: 1
    PublicTag:
      description: "公开标签资源。"
      type: object
      additionalProperties: false
      required: [id, name, articleCount]
      properties:
        id:
          $ref: '#/components/schemas/PositiveID'
        name:
          description: "标签名称。"
          type: string
          minLength: 1
          maxLength: 60
        articleCount:
          description: "满足公开条件的文章数量。"
          type: integer
          format: int64
          minimum: 1
    PublicArticleTypeList:
      description: "公开文章类型分页响应信封。"
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          description: "当前页公开文章类型数组。"
          type: array
          items:
            $ref: '#/components/schemas/PublicArticleType'
        page:
          $ref: '#/components/schemas/PageInfo'
    PublicTagList:
      description: "公开标签分页响应信封。"
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          description: "当前页公开标签数组。"
          type: array
          items:
            $ref: '#/components/schemas/PublicTag'
        page:
          $ref: '#/components/schemas/PageInfo'
    Article:
      description: "文章资源，包含所属类型、标签关系、状态、计数、版本和 UTC 时间。"
      type: object
      additionalProperties: false
      required: [id, title, slug, digest, content, articleTypeId, articleType, tagIds, status, support, comment, visited, version, createdAt, updatedAt]
      properties:
        title:
          description: "标题。"
          type: string
          minLength: 1
          maxLength: 120
        slug:
          description: "规范短标识，例如 `hello-world`。"
          type: string
          minLength: 1
          maxLength: 160
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        digest:
          description: "文章摘要。"
          type: string
          minLength: 1
          maxLength: 500
        content:
          description: "文章正文。"
          type: string
          minLength: 1
        id:
          $ref: '#/components/schemas/PositiveID'
        articleTypeId:
          $ref: '#/components/schemas/PositiveID'
        articleType:
          $ref: '#/components/schemas/PublicArticleTypeSummary'
        tagIds:
          description: "关联标签正整数标识数组，值不重复。"
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/PositiveID'
        status:
          $ref: '#/components/schemas/ArticleStatus'
        support:
          description: "获赞次数，非负。"
          type: integer
          format: int64
          minimum: 0
        comment:
          description: "评论次数，非负。"
          type: integer
          format: int64
          minimum: 0
        visited:
          description: "访问次数，非负。"
          type: integer
          format: int64
          minimum: 0
        version:
          $ref: '#/components/schemas/Version'
        createdAt:
          description: "UTC 创建时间，例如 `2026-01-01T00:00:00Z`。"
          type: string
          format: date-time
        updatedAt:
          description: "UTC 更新时间，从未更新时为 `null`。"
          type: string
          format: date-time
          nullable: true
...
    ArticleList:
      description: "文章分页响应信封。"
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          description: "当前页资源数组。"
          type: array
          items:
            $ref: '#/components/schemas/Article'
        page:
          $ref: '#/components/schemas/PageInfo'
...